Refactoring routes, adding middleware.
Using a body with GET requests is a little unorthodox, but I feel like
it presents a much better way to input and consume the information
needed by Genesis.

I've got a middleware that injects a Map value into the context, so now
I only have to write the loading code once. Eventually, the middleware
can do validations and return helpful errors.